DX: Simplify tracking of resources
There's an assumption that the resources are created in the same order as they appear in the descriptor heap, which is the order they appear when traversing Pipeline/Sets/Resources. The root signature is where the mapping from index in the descriptor heap to a shader register happens.

Because of this, trying to keep track of the resources via their DX register binding makes things more complicated that necessary, since we can just assume there's a simple mapping of Set/Resource to the resources.