Introduce a simpler StructuredBuffer test
This moves the current structured buffer test to a new file and
introduces a simpler one that doesn't need llvm/llvm-project#104503 and
llvm/llvm-project#121010 to be in place in order to pass.

Note: this would be better if it exercised both SRVs and UAVs, but SRV
support isn't implemented in the test framework yet.