Add color utility
This PR adds a new color utility object for working with colors and transforming color spaces. It currently supports translating between RGB, CIEXYZ and CIELAB. The D65 white point is hard coded, but could be refactored to be configurable.

This change also adds unit testing for the Image library so that this can be tested more cleanly, and adds the unit test target to the test-all build configuration.