Add new imagediff tool
This change introduces a new image diff tool that has two modes for comparing PNG images.

The "Exact Match" mode loads a PNG and compares the decoded image data. This comensates for differences in the image generation tool, image headers or other encoding differences.

The "CIE76" mode compares the two images computing the CIE76 color difference, and fails if the difference of the worst pixel is greater than 2.3, which cooresponds to a "just noticable" difference.

The tool can also generate an image that represents the difference between the two rendered images.

In addition to introducing this tooling, this change also enables golden-image verification testing. Golden images are stored in a separate git repo (https://github.com/llvm-beanz/offload-golden-images), and are opted into during configuring the build by setting `GOLDENIMAGE_DIR` to the root of that checkout.

The only image currently tested is the Mandelbrot test.